feat(runtimes): extract reusable runtime logic into the `base-runtime.ts`
